requirements_minimal.txt
# Core dependencies
cdp-sdk @ git+https://github.com/coinbase/cdp-sdk-python.git@v0.0.2
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-vertexai
langchain-google-genai
langchain-community>=0.3.0
mem0ai
langsmith
google-generativeai

# API
fastapi
uvicorn[standard]
websockets
httpx

# Google Cloud
google-cloud-firestore
google-cloud-secret-manager

# Utilities
python-dotenv
pydantic-settings
pandas
numpy
aiohttp

# Development
pytest
pytest-asyncio